Extract org_cache from auth_cache
